build: enable `Swagger` configuration
