+++ build.gradle.kts
@@ -1,13 +1,12 @@
 plugins {
     application
-    java
     id("org.springframework.boot") version "3.3.5"
     id("io.spring.dependency-management") version "1.1.6"
     id("org.sonarqube") version "5.1.0.4882"
     id("jacoco")
 }
 
-group = "com.softdevsix.argos"
+group = "com.softdevsix"
 version = "0.0.1-SNAPSHOT"
 
 java {
@@ -43,11 +42,10 @@ val mockitoDocVersion = "5.5.0"
 
 dependencies {
     implementation("org.springframework.boot:spring-boot-starter-data-jpa")
-    implementation("org.springframework.boot:spring-boot-starter-data-rest")
     implementation("org.springframework.boot:spring-boot-starter-web")
-    implementation("org.springdoc:springdoc-openapi-starter-webmvc-ui:$springDocVersion")
     implementation("org.flywaydb:flyway-core")
     implementation("org.flywaydb:flyway-database-postgresql")
+    implementation("org.springdoc:springdoc-openapi-starter-webmvc-ui:$springDocVersion")
     compileOnly("org.projectlombok:lombok")
     developmentOnly("org.springframework.boot:spring-boot-devtools")
     runtimeOnly("org.postgresql:postgresql")
@@ -58,13 +56,12 @@ dependencies {
 }
 
 application {
-    mainClass.set("com.softdevsix.api.ArgosApplication")
+    mainClass.set("com.softdevsix.ArgosApplication")
 }
 
 tasks.withType<Test> {
     useJUnitPlatform()
 }
-
 
 jacoco {
     toolVersion = "0.8.8"
