--- src/main/java/com/softdevsix/api/ArgosApplication.java
@@ -1,13 +1,0 @@
-package com.softdevsix.api;
-
-import org.springframework.boot.SpringApplication;
-import org.springframework.boot.autoconfigure.SpringBootApplication;
-
-@SpringBootApplication
-public class ArgosApplication {
-
-    public static void main(String[] args) {
-        SpringApplication.run(ArgosApplication.class, args);
-    }
-
-}
