--- src/main/java/com/softdevsix/api/configurations/SwaggerConfig.java
@@ -1,19 +1,0 @@
-package com.softdevsix.api.configurations;
-
-import io.swagger.v3.oas.models.OpenAPI;
-import io.swagger.v3.oas.models.info.Info;
-import org.springframework.context.annotation.Bean;
-import org.springframework.context.annotation.Configuration;
-
-@Configuration
-public class SwaggerConfig {
-
-    @Bean
-    public OpenAPI customOpenAPI() {
-        return new OpenAPI()
-                .info(new Info()
-                        .title("Argos API")
-                        .version("1.0")
-                        .description("An application for verifying code coverage and file coverage in project."));
-    }
-}
