--- src/main/java/com/softdevsix/api/controllers/Deleteme.md
@@ -1,1 +1,0 @@
-Delete
