--- src/main/java/com/softdevsix/api/domain/entities/file/FileCoverageResult.java
@@ -1,17 +1,0 @@
-package com.softdevsix.api.domain.entities.file;
-
-import jakarta.persistence.GeneratedValue;
-import jakarta.persistence.GenerationType;
-import lombok.Builder;
-import lombok.Data;
-
-import java.util.List;
-import java.util.UUID;
-
-@Data
-@Builder
-public class FileCoverageResult {
-    @GeneratedValue(strategy = GenerationType.UUID)
-    private UUID id;
-    private List<MethodCoverageResult> allStatements;
-}
