--- src/main/java/com/softdevsix/api/exceptions/FileNotFoundException.java
@@ -1,8 +1,0 @@
-package com.softdevsix.api.exceptions;
-
-public class FileNotFoundException extends Exception {
-
-    public FileNotFoundException(String message) {
-        super(message);
-    }
-}
