--- src/main/java/com/softdevsix/api/exceptions/ProjectNotFoundException.java
@@ -1,4 +1,0 @@
-package com.softdevsix.api.exceptions;
-
-public class ProjectNotFoundException extends Exception{
-}
