--- src/main/java/com/softdevsix/api/mappers/requests/deleteme.md
@@ -1,1 +1,0 @@
-dele
