--- src/main/java/com/softdevsix/api/mappers/responses/deleteme.md
@@ -1,1 +1,0 @@
-dele
