--- src/main/java/com/softdevsix/api/repositories/IProjectRepository.java
@@ -1,4 +1,0 @@
-package com.softdevsix.api.repositories;
-
-public interface IProjectRepository {
-}
