--- src/main/java/com/softdevsix/api/services/IProjectService.java
@@ -1,4 +1,0 @@
-package com.softdevsix.api.services;
-
-public interface IProjectService {
-}
