+++ src/main/java/com/softdevsix/application/mapper/requests/ProjectCreateRequest.java
@@ -1,0 +1,4 @@
+package com.softdevsix.application.mapper.requests;
+
+public class ProjectCreateRequest {
+}
