+++ src/main/java/com/softdevsix/application/mapper/responses/ProjectResponse.java
@@ -1,0 +1,4 @@
+package com.softdevsix.application.mapper.responses;
+
+public class ProjectResponse {
+}
