+++ src/main/java/com/softdevsix/application/service/IProjectService.java
@@ -1,0 +1,4 @@
+package com.softdevsix.application.service;
+
+public interface IProjectService {
+}
