+++ src/main/java/com/softdevsix/application/service/ProjectService.java
@@ -1,0 +1,4 @@
+package com.softdevsix.application.service;
+
+public class ProjectService {
+}
