+++ src/main/java/com/softdevsix/controller/ProjectController.java
@@ -1,0 +1,4 @@
+package com.softdevsix.controller;
+
+public class ProjectController {
+}
