--- src/main/java/com/softdevsix/controllers/ProjectController.java
@@ -1,4 +1,0 @@
-package com.softdevsix.controllers;
-
-public class ProjectController {
-}
