--- src/main/java/com/softdevsix/domain/entities/file/File.java
@@ -1,4 +1,0 @@
-package com.softdevsix.domain.entities.file;
-
-public class File {
-}
