--- src/main/java/com/softdevsix/domain/entities/project/Project.java
@@ -1,22 +1,0 @@
-package com.softdevsix.domain.entities.project;
-
-import com.softdevsix.domain.entities.file.File;
-import jakarta.persistence.GeneratedValue;
-import jakarta.persistence.GenerationType;
-import lombok.Builder;
-import lombok.Data;
-
-import java.util.List;
-import java.util.UUID;
-
-@Data
-@Builder
-public class Project {
-    @GeneratedValue(strategy = GenerationType.UUID)
-    private UUID projectId;
-    private String name;
-    private String description;
-    private ProjectParams projectParams;
-    private ProjectResults projectResults;
-    private List<File> coveredFiles;
-}
