--- src/main/java/com/softdevsix/domain/entities/project/ProjectParams.java
@@ -1,17 +1,0 @@
-package com.softdevsix.domain.entities.project;
-
-import jakarta.persistence.GeneratedValue;
-import jakarta.persistence.GenerationType;
-import lombok.Builder;
-import lombok.Data;
-
-import java.util.UUID;
-
-@Data
-@Builder
-public class ProjectParams {
-    @GeneratedValue(strategy = GenerationType.UUID)
-    private UUID id;
-    private float requiredCoveragePercentage;
-    private char requiredCodeRating;
-}
