--- src/main/java/com/softdevsix/domain/entities/project/ProjectResults.java
@@ -1,21 +1,0 @@
-package com.softdevsix.domain.entities.project;
-
-import com.softdevsix.domain.coverage.ProjectCoverageResult;
-import com.softdevsix.domain.staticanalysis.CodeAnalysisResult;
-import jakarta.persistence.GeneratedValue;
-import jakarta.persistence.GenerationType;
-import lombok.Builder;
-import lombok.Data;
-
-import java.util.UUID;
-
-@Data
-@Builder
-public class ProjectResults {
-    @GeneratedValue(strategy = GenerationType.UUID)
-    private UUID id;
-    private UUID projectId;
-    private Status status;
-    private ProjectCoverageResult coverageResult;
-    private CodeAnalysisResult codeAnalysisResult;
-}
