--- src/main/java/com/softdevsix/domain/entities/project/Status.java
@@ -1,6 +1,0 @@
-package com.softdevsix.domain.entities.project;
-
-public enum Status {
-    PASSED,
-    FAILED,
-}
