+++ src/main/java/com/softdevsix/domain/entity/file/File.java
@@ -1,0 +1,4 @@
+package com.softdevsix.domain.entity.file;
+
+public class File {
+}
