+++ src/main/java/com/softdevsix/domain/entity/file/FileCoverageResult.java
@@ -1,0 +1,17 @@
+package com.softdevsix.domain.entity.file;
+
+import jakarta.persistence.GeneratedValue;
+import jakarta.persistence.GenerationType;
+import lombok.Builder;
+import lombok.Data;
+
+import java.util.List;
+import java.util.UUID;
+
+@Data
+@Builder
+public class FileCoverageResult {
+    @GeneratedValue(strategy = GenerationType.UUID)
+    private UUID id;
+    private List<MethodCoverageResult> allStatements;
+}
