+++ src/main/java/com/softdevsix/domain/entity/project/Status.java
@@ -1,0 +1,6 @@
+package com.softdevsix.domain.entity.project;
+
+public enum Status {
+    PASSED,
+    FAILED,
+}
