+++ src/main/java/com/softdevsix/domain/entity/staticanalysis/Rating.java
@@ -1,0 +1,8 @@
+package com.softdevsix.domain.entity.staticanalysis;
+
+public enum Rating {
+    A,
+    B,
+    C,
+    D,
+}
