+++ src/main/java/com/softdevsix/domain/exception/FileNotFoundException.java
@@ -1,0 +1,8 @@
+package com.softdevsix.domain.exception;
+
+public class FileNotFoundException extends Exception {
+
+    public FileNotFoundException(String message) {
+        super(message);
+    }
+}
