+++ src/main/java/com/softdevsix/domain/exception/ProjectNotFoundException.java
@@ -1,0 +1,4 @@
+package com.softdevsix.domain.exception;
+
+public class ProjectNotFoundException {
+}
