+++ src/main/java/com/softdevsix/domain/repository/IFileRepository.java
@@ -1,0 +1,9 @@
+package com.softdevsix.domain.repository;
+
+import com.softdevsix.api.domain.entities.file.File;
+import org.springframework.data.jpa.repository.JpaRepository;
+
+import java.util.UUID;
+
+public interface IFileRepository extends JpaRepository<File, UUID> {
+}
