+++ src/main/java/com/softdevsix/domain/repository/IProjectRepository.java
@@ -1,0 +1,4 @@
+package com.softdevsix.domain.repository;
+
+public interface IProjectRepository {
+}
