--- src/main/java/com/softdevsix/domain/staticanalysis/Rating.java
@@ -1,8 +1,0 @@
-package com.softdevsix.domain.staticanalysis;
-
-public enum Rating {
-    A,
-    B,
-    C,
-    D,
-}
