--- src/main/java/com/softdevsix/exceptions/ProjectNotFoundException.java
@@ -1,4 +1,0 @@
-package com.softdevsix.exceptions;
-
-public class ProjectNotFoundException {
-}
