--- src/main/java/com/softdevsix/mappers/requests/ProjectCreateRequest.java
@@ -1,4 +1,0 @@
-package com.softdevsix.mappers.requests;
-
-public class ProjectCreateRequest {
-}
