--- src/main/java/com/softdevsix/mappers/responses/ProjectResponse.java
@@ -1,4 +1,0 @@
-package com.softdevsix.mappers.responses;
-
-public class ProjectResponse {
-}
