--- src/main/java/com/softdevsix/repositories/IProjectRepository.java
@@ -1,4 +1,0 @@
-package com.softdevsix.repositories;
-
-public interface IProjectRepository {
-}
