--- src/main/java/com/softdevsix/services/IProjectService.java
@@ -1,4 +1,0 @@
-package com.softdevsix.services;
-
-public interface IProjectService {
-}
