--- src/main/java/com/softdevsix/services/ProjectService.java
@@ -1,4 +1,0 @@
-package com.softdevsix.services;
-
-public class ProjectService {
-}
